muncipal_backend/pyproject.toml
[project]
name = "muncipal-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi>=0.136.1",
    "fastapi-users[sqlalchemy]>=15.0.5",
    "joblib>=1.5.3",
    "pandas>=3.0.2",
    "psycopg2-binary>=2.9.12",
    "python-dotenv>=1.2.2",
    "scikit-learn==1.7.2",
    "sqlalchemy[asyncio]>=2.0.49",
    "uvicorn[standard]>=0.46.0",
]
